refactor(性能测试): 重构性能测试解析jmx的方式
