+++ backend/src/main/java/io/metersphere/performance/parse/EngineSourceParser.java
@@ -5,5 +5,5 @@ import io.metersphere.performance.engine.EngineContext;
 import java.io.InputStream;
 
 public interface EngineSourceParser {
-    String parse(EngineContext context, InputStream source) throws Exception;
+    byte[] parse(EngineContext context, InputStream source) throws Exception;
 }
